+++ src/mixer.h
@@ -10,8 +10,8 @@ uint8_t mixerr_percentage(struct text_object *obj);
 int check_mixer_muted(struct text_object *);
 
 void scan_mixer_bar(struct text_object *, const char *);
-uint8_t mixer_barval(struct text_object *);
-uint8_t mixerl_barval(struct text_object *);
-uint8_t mixerr_barval(struct text_object *);
+double mixer_barval(struct text_object *);
+double mixerl_barval(struct text_object *);
+double mixerr_barval(struct text_object *);
 
 #endif /*MIXER_H_*/
